create separate edit time block component.
change 'Event' interface name -> CalendarEvent to avoid conflicts with pre-defined DOM type
